Delete Realloc
The current implementation is not desirable, and Realloc is not an
important feature, especially at this stage.

We should avoid touching the shared memory pages in the daemon (except
for debugging purposes, such as pattern-filling). If implementing
Realloc in the future, we should take full advantage of a custom
allocator so that the object can be extended to the left as well as to
the right, and do any memcpy() or memmove() in the client.