Delete the 'clear' flag in requests
We want to move toward the daemon not touching shared memory pages by
default. Clearing should be done by clients except in cases where the
daemon allocates fresh pages that happen to be zero-filled.

Added a response flag to the protocol, although we will only use it in
the future as an optimization.