Use local wrap for flatcc
Although flatcc had a 'meson' branch with a few releases, it has not
been maintained (last was meson-0.5.3; current trunk is 0.6.1).

The CMake build of flatcc is broken (places build artifacts in the
source directory) and does not work with Meson's cmake module.

Installing flatcc is also likely to not work well, because it does not
provide a pkg-config .pc file.

So the simplest thing to do is to do our own minimal Meson build. It is
quite simple when all the tests and options are left out.