Revise usage of 'noexcept'
Even though we do not use exceptions to handle recoverable errors,
scattering 'noexcept' everywhere does not offer any benefit. Instead, it
may get in the way of future unit testing of contract checks.

Also, it probably make sense to limit compiler-enforcement of no-throw
(i.e., calling std::terminate() in the event of an exception) to
destructors and move assignment operators.

'noexcept' should generally be limited to wide-contract functions (which
are relatively scarse), move constructors and assignment operators,
swap(), and (implicitly) destructors (see the so-called Lakos Rule).